third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-basic-non-simple-deny-cached-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-non-get-allow.cgi. Request header field Content-Type is not allowed by Access-Control-Allow-Headers.
PASS: Exception thrown. Cross-domain access is not allowed in 'send'. [Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/access-control-basic-non-get-allow.cgi'.].
